fix ctrl-c issue
